Moves docs/ to root folder
The GitHub pages site for this project was served
from docs/ but then the test files couldn’t link
to create-mail.js, which is the whole point. So
this commit removes docs/. It moves /docs/tests/
to /tests, and deletes /docs/index.md which is
only a copy of /README.md anyways, so nothing lost.